refactor: move seedPhrase getter inside useWallets
